testing on prototype complete for 4 limit switches to control wheel spin in dignostic limit switches mode

+++ trunk/src/team1100/season2010/robot/ChainRotationMotor.java
@@ -70,7 +70,7 @@ public class ChainRotationMotor
           pot_deadband = potDeadband;
       }
 
-      public double getPot()
+      public int getPot()
       {
           return pot.getAverageValue();
       }
